+++ service/src/main/java/de/starwit/service/ServiceConfiguration.java
@@ -19,11 +19,13 @@ import org.springframework.data.redis.connection.stream.ReadOffset;
 import org.springframework.data.redis.connection.stream.StreamOffset;
 import org.springframework.data.redis.stream.StreamMessageListenerContainer;
 import org.springframework.data.redis.stream.Subscription;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 import io.lettuce.core.ClientOptions;
 import io.lettuce.core.ClientOptions.DisconnectedBehavior;
 
 @SpringBootApplication
+@EnableScheduling
 @ComponentScan(basePackages = "de.starwit.persistence")
 public class ServiceConfiguration {